+++ CHANGELOG.md
@@ -4,7 +4,25 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## UNRELEASED
 
+## TODO
+- Logs
+  - Save: User, Title, Subreddit, Post Url, Link Url, Date Done, Path Saved
 
+### MODFIED
+- cachemysubreddit/
+  - Renamed to cms/
+### REWRITTEN
+- cms/imgur.py
+  - Changed to requests only api, less likey to run into port issues.
+- cms/config.py
+  - Changed to an .ini based config
+- cms/main.py
+  - Removed subreddit functionality for this release
+- cms/reddit.py
+  - Changed to use praw-4.0 
+  - Refactored submitted code
+
+## [0.5.0] - 2016-12-24
 ## [0.4.1] - 2016-08-20
 ### MODIFIED
 - cachemysubreddit/imgur.py